requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
google-generativeai
